Refactor versioning system: remove versioning.js, introduce versioning.ts; update related scripts and components
- Deleted public/versioning.js and migrated its functionality to src/services/versioning.ts.
- Updated version bumping script to reflect new versioning file location.
- Adjusted version comparison logic in various services to utilize the new versioning module.
- Introduced new reusable web components: fill-box and slider-input.
- Added heightmap templates and pre-created heightmaps data structures.
- Updated index.html to load new components and removed obsolete scripts.
- Cleaned up global variable declarations in types/global.ts.
