+++ public/sw.js
@@ -20,7 +20,6 @@ registerRoute(
   ({request, url}) =>
     request.destination === "script" &&
     !url.pathname.endsWith("min.js") &&
-    !url.pathname.includes("versioning.js") &&
     !url.pathname.includes("google"),
   new StaleWhileRevalidate({
     cacheName: "fmg-scripts",
