+++ src/components/index.ts
@@ -1,0 +1,3 @@
+// Reusable web components; importing registers the custom elements
+import "./fill-box";
+import "./slider-input";
