+++ src/index.html
@@ -5509,24 +5509,22 @@
         <g id="trade-markers"></g>
       </defs>
     </svg>
-    <script src="components/fill-box.js"></script>
-    <script src="components/slider-input.js"></script>
     <script src="libs/jquery-3.1.1.min.js"></script>
     <script src="libs/jquery-ui.min.js"></script>
-    <script src="versioning.js"></script>
     <script src="libs/d3.min.js"></script>
     <script src="libs/flatqueue.js"></script>
     <script src="libs/delaunator.min.js"></script>
     <script src="libs/indexedDB.js?v=1.99.00"></script>
+    <script type="module" src="components/index.ts"></script>
     <script type="module" src="data/supporters.ts"></script>
     <script type="module" src="data/view-3d-options.ts"></script>
+    <script type="module" src="data/heightmap-templates.ts"></script>
+    <script type="module" src="data/precreated-heightmaps.ts"></script>
     <script type="module" src="utils/index.ts"></script>
     <script type="module" src="generators/index.ts"></script>
     <script type="module" src="renderers/index.ts"></script>
     <script type="module" src="controllers/index.ts"></script>
     <script type="module" src="services/index.ts"></script>
-    <script defer src="config/heightmap-templates.js"></script>
-    <script defer src="config/precreated-heightmaps.js"></script>
     <script defer src="libs/alea.min.js?v1.105.0"></script>
     <script defer src="libs/polylabel.min.js?v1.105.0"></script>
     <script defer src="libs/simplify.js?v1.105.6"></script>
