+++ src/services/index.ts
@@ -1,6 +1,7 @@
 import { createRegistry } from "@/utils/registry";
 import "./autosave";
 import "./fonts";
+import "./versioning";
 
 export const Services = createRegistry({
   Cloud: () => import("@/services/io/cloud").then(m => m.CloudStorage),
