+++ src/services/io/load.ts
@@ -1,5 +1,6 @@
 import { select } from "d3";
 import { Services } from "@/services";
+import { cleanupData, compareVersions, isValidVersion, parseMapVersion, VERSION } from "@/services/versioning";
 import { calculateVoronoi, ensureEl, last, link, minmax, parseError, rn } from "@/utils";
 
 async function quickLoad(): Promise<void> {
